Rename `sample.py` to `samplers.py`
