+++ docs/api/samplers.md
@@ -1,0 +1,1 @@
+::: outlines.text.generate.samplers
