+++ tests/text/generate/test_samplers.py
@@ -2,7 +2,11 @@ import math
 
 import torch
 
-from outlines.text.generate.sample import greedy, multinomial, vectorized_random_choice
+from outlines.text.generate.samplers import (
+    greedy,
+    multinomial,
+    vectorized_random_choice,
+)
 
 
 def test_greedy():
